Add: useVelcro fetching logic hook
